requirements.txt
Package             Version
------------------- ------------
anyio               3.6.2
APScheduler         3.6.3
cached-property     1.5.2
cachetools          4.2.2
certifi             2023.5.7
cffi                1.15.1
charset-normalizer  3.2.0
cryptography        41.0.1
decorator           5.1.1
filelock            3.12.2
greenlet            2.0.2
h11                 0.14.0
httpcore            0.17.0
httpx               0.24.0
idna                3.4
inflection          0.5.1
Jinja2              3.1.2
MarkupSafe          2.1.3
mpmath              1.3.0
mypy-extensions     1.0.0
networkx            3.1
numpy               1.25.0
Pillow              10.0.0
pip                 22.3.1
pycparser           2.21
python-telegram-bot 20.3
pytz                2023.3
requests            2.31.0
setuptools          65.5.1
six                 1.16.0
sniffio             1.3.0
SQLAlchemy          2.0.4
sqlalchemy-orm      1.2.4
sympy               1.12
torch               2.0.1
torchvision         0.15.2
tornado             6.3.2
typing_extensions   4.5.0
typing-inspect      0.8.0
tzdata              2023.3
tzlocal             5.0.1
urllib3             2.0.3
wheel               0.38.4
